refactor(frontend): Consolidate preferences into single key and merge storage modules
Consolidate 8 individual preference localStorage keys into a single
leapmux:browser-prefs JSON object, mirroring the backend's
storedPreferences schema. Absent fields mean "use account default",
replacing the old 'account-default' sentinel value.

- Merge safeStorage.ts and storageCleanup.ts into browserStorage.ts
  as a single module for all localStorage concerns
- Replace KEY_THEME, KEY_TERMINAL_THEME, KEY_DIFF_VIEW, etc. with
  single KEY_BROWSER_PREFS storing a BrowserPreferences JSON object
- Move enterKeyMode from draftManagement.ts into PreferencesContext
  as part of the consolidated preferences
- Update cross-tab theme sync in app.tsx to diff the browser-prefs
  object on StorageEvent
- Add getBrowserPref/setInitialBrowserPref E2E test helpers
- Update all E2E tests for consolidated storage format
