+++ frontend/src/app.tsx
@@ -9,9 +9,8 @@ import { LauncherView } from '~/components/desktop/LauncherView'
 import { UserMenuDialogs } from '~/components/shell/UserMenu'
 import { AuthProvider } from '~/context/AuthContext'
 import { PreferencesProvider, usePreferences } from '~/context/PreferencesContext'
+import { initStorageCleanup, KEY_BROWSER_PREFS, loadBrowserPrefs } from '~/lib/browserStorage'
 import { resolveStack } from '~/lib/resolveStack'
-import { safeGetString } from '~/lib/safeStorage'
-import { initStorageCleanup, KEY_THEME } from '~/lib/storageCleanup'
 import { disableTextSubstitutions } from '~/lib/textInputBehavior'
 import { heightFull } from '~/styles/shared.css'
 import '~/lib/oat'
@@ -24,10 +23,10 @@ export type ThemePreference = 'light' | 'dark' | 'system'
 
 /** Read the saved theme preference from localStorage (instant, no flash). */
 function getStoredTheme(): ThemePreference {
-  const stored = safeGetString(KEY_THEME)
+  const stored = loadBrowserPrefs().theme
   if (stored === 'light' || stored === 'dark' || stored === 'system')
     return stored
-  return 'system' // 'account-default' or missing → default until account loads
+  return 'system' // missing → default until account loads
 }
 
 /** Resolve the effective theme based on preference + system setting. */
@@ -156,12 +155,16 @@ export default function App() {
 
   // Listen for localStorage changes from other tabs.
   const handleStorage = (e: StorageEvent) => {
-    if (e.key === KEY_THEME) {
-      const val = e.newValue
-      if (val === 'light' || val === 'dark' || val === 'system')
-        setThemePreference(val)
-      else
-        setThemePreference('system') // 'account-default' or null
+    if (e.key === KEY_BROWSER_PREFS) {
+      const oldPrefs = e.oldValue ? JSON.parse(e.oldValue) : {}
+      const newPrefs = e.newValue ? JSON.parse(e.newValue) : {}
+      if (oldPrefs.theme !== newPrefs.theme) {
+        const val = newPrefs.theme
+        if (val === 'light' || val === 'dark' || val === 'system')
+          setThemePreference(val)
+        else
+          setThemePreference('system')
+      }
     }
   }
   window.addEventListener('storage', handleStorage)
