+++ frontend/src/components/chat/AgentEditorPanel.tsx
@@ -16,9 +16,8 @@ import { showWarnToast } from '~/components/common/Toast'
 import { Tooltip } from '~/components/common/Tooltip'
 import { AgentProvider } from '~/generated/leapmux/v1/agent_pb'
 import { createLoadingSignal } from '~/hooks/createLoadingSignal'
+import { PREFIX_EDITOR_MIN_HEIGHT, safeGetString, safeRemoveItem, safeSetString } from '~/lib/browserStorage'
 import { formatResetTimestamp, getResetsAt } from '~/lib/rateLimitUtils'
-import { safeGetString, safeRemoveItem, safeSetString } from '~/lib/safeStorage'
-import { PREFIX_EDITOR_MIN_HEIGHT } from '~/lib/storageCleanup'
 import { registerEditorRef, unregisterEditorRef } from '~/stores/editorRef.store'
 import { spinner } from '~/styles/animations.css'
 import { iconSize } from '~/styles/tokens'
