+++ frontend/src/components/chat/controlResponseHandling.ts
@@ -5,9 +5,8 @@ import type { ControlRequest } from '~/stores/control.store'
 import type { PermissionMode } from '~/utils/controlResponse'
 import { createEffect, createMemo, on } from 'solid-js'
 import { AgentProvider } from '~/generated/leapmux/v1/agent_pb'
+import { PREFIX_ASK_STATE, safeGetJson, safeRemoveItem, safeSetJson } from '~/lib/browserStorage'
 import { clearDraft } from '~/lib/editor/draftPersistence'
-import { safeGetJson, safeRemoveItem, safeSetJson } from '~/lib/safeStorage'
-import { PREFIX_ASK_STATE } from '~/lib/storageCleanup'
 import { buildAllowResponse, buildDenyResponse, getToolInput, getToolName } from '~/utils/controlResponse'
 import { buildAskAnswers, trySubmitAskUserQuestion } from './controls/AskUserQuestionControl'
 import { sendCodexUserInputResponse } from './controls/CodexControlRequest'
