+++ frontend/src/components/tree/FilesSection.tsx
@@ -14,7 +14,7 @@ import { Icon } from '~/components/common/Icon'
 import { IconButton, IconButtonState } from '~/components/common/IconButton'
 import { RefreshButton } from '~/components/common/RefreshButton'
 import { GitFileStatusCode } from '~/generated/leapmux/v1/common_pb'
-import { safeGetJson, safeSetJson } from '~/lib/safeStorage'
+import { safeGetJson, safeSetJson } from '~/lib/browserStorage'
 import { DirectoryTree } from './DirectoryTree'
 import * as styles from './FilesSection.css'
 import { DiffStatsBadge, getGitFileIconClass } from './gitStatusUtils'
