+++ frontend/src/context/PreferencesContext.tsx
@@ -1,23 +1,17 @@
 import type { ParentComponent } from 'solid-js'
 import type { ThemePreference } from '~/app'
+import type { BrowserPreferences, EnterKeyMode } from '~/lib/browserStorage'
 import type { TerminalThemePreference } from '~/lib/terminal'
 import { createContext, createEffect, createSignal, onMount, useContext } from 'solid-js'
 import { userClient } from '~/api/clients'
 import { DiffView, TurnEndSound } from '~/generated/leapmux/v1/user_pb'
+import { KEY_BROWSER_PREFS, loadBrowserPrefs, safeSetJson } from '~/lib/browserStorage'
 import { setDebugEnabled } from '~/lib/logger'
-import { safeGetString, safeSetString } from '~/lib/safeStorage'
-import { KEY_DEBUG_LOGGING, KEY_DIFF_VIEW, KEY_SHOW_HIDDEN_MESSAGES, KEY_TERMINAL_THEME, KEY_THEME, KEY_TURN_END_SOUND, KEY_TURN_END_SOUND_VOLUME } from '~/lib/storageCleanup'
 
 const DEFAULT_MONO_FONT_FAMILY = '"Hack NF", Hack, "SF Mono", Consolas, monospace'
 
 export type DiffViewPreference = 'unified' | 'split'
 export type TurnEndSoundPreference = 'none' | 'ding-dong'
-
-/**
- * Sentinel value stored in localStorage to indicate
- * "use the account default" for this browser setting.
- */
-const ACCOUNT_DEFAULT = 'account-default'
 
 interface PreferencesState {
   /** Resolved theme preference (localStorage override → account default → hardcoded default). */
@@ -47,6 +41,9 @@ interface PreferencesState {
   /** Whether hidden messages are shown in the chat view (developer feature). */
   showHiddenMessages: () => boolean
   setShowHiddenMessages: (value: boolean) => void
+  /** Resolved enter key mode. */
+  enterKeyMode: () => EnterKeyMode
+  setEnterKeyMode: (value: EnterKeyMode) => void
 
   // --- Browser-level overrides (localStorage) ---
   /** Raw browser-level theme override. null means "use account default". */
@@ -113,20 +110,16 @@ function turnEndSoundToProto(tes: TurnEndSoundPreference): TurnEndSound {
   return tes === 'ding-dong' ? TurnEndSound.DING_DONG : TurnEndSound.NONE
 }
 
-function readLocalStorage(key: string): string | null {
-  const v = safeGetString(key)
-  if (v === ACCOUNT_DEFAULT || v === null || v === '')
-    return null
-  return v
-}
-
-function writeLocalStorage(key: string, value: string | null) {
-  if (value === null) {
-    safeSetString(key, ACCOUNT_DEFAULT)
+/** Update a single field in the consolidated browser preferences. */
+function updateBrowserPref(key: keyof BrowserPreferences, value: BrowserPreferences[keyof BrowserPreferences] | undefined): void {
+  const prefs = loadBrowserPrefs()
+  if (value === undefined) {
+    delete prefs[key]
   }
   else {
-    safeSetString(key, value)
+    (prefs as Record<string, unknown>)[key] = value
   }
+  safeSetJson(KEY_BROWSER_PREFS, prefs)
 }
 
 export const PreferencesProvider: ParentComponent = (props) => {
@@ -142,17 +135,19 @@ export const PreferencesProvider: ParentComponent = (props) => {
   const [accountTurnEndSoundVolume, setAccountTurnEndSoundVolume] = createSignal<number>(100)
   const [accountDebugLogging, setAccountDebugLogging] = createSignal(false)
 
-  // --- Browser-level (localStorage) ---
+  // --- Browser-level (localStorage) --- load once from consolidated key
+  const initialPrefs = loadBrowserPrefs()
+
   const [browserTheme, setBrowserThemeSignal] = createSignal<ThemePreference | null>(
-    readLocalStorage(KEY_THEME) as ThemePreference | null,
+    (initialPrefs.theme as ThemePreference) ?? null,
   )
   const [browserTerminalTheme, setBrowserTerminalThemeSignal] = createSignal<TerminalThemePreference | null>(
-    readLocalStorage(KEY_TERMINAL_THEME) as TerminalThemePreference | null,
+    (initialPrefs.terminalTheme as TerminalThemePreference) ?? null,
   )
 
   const setBrowserTheme = (value: ThemePreference | null) => {
     setBrowserThemeSignal(value)
-    writeLocalStorage(KEY_THEME, value)
+    updateBrowserPref('theme', value ?? undefined)
     // Notify app.tsx for instant reactivity
     const setter = (window as any).__leapmux_setTheme
     if (setter) {
@@ -162,60 +157,60 @@ export const PreferencesProvider: ParentComponent = (props) => {
 
   const setBrowserTerminalTheme = (value: TerminalThemePreference | null) => {
     setBrowserTerminalThemeSignal(value)
-    writeLocalStorage(KEY_TERMINAL_THEME, value)
+    updateBrowserPref('terminalTheme', value ?? undefined)
   }
 
   const [browserDiffView, setBrowserDiffViewSignal] = createSignal<DiffViewPreference | null>(
-    readLocalStorage(KEY_DIFF_VIEW) as DiffViewPreference | null,
+    (initialPrefs.diffView as DiffViewPreference) ?? null,
   )
 
   const setBrowserDiffView = (value: DiffViewPreference | null) => {
     setBrowserDiffViewSignal(value)
-    writeLocalStorage(KEY_DIFF_VIEW, value)
+    updateBrowserPref('diffView', value ?? undefined)
   }
 
   const [browserTurnEndSound, setBrowserTurnEndSoundSignal] = createSignal<TurnEndSoundPreference | null>(
-    readLocalStorage(KEY_TURN_END_SOUND) as TurnEndSoundPreference | null,
+    (initialPrefs.turnEndSound as TurnEndSoundPreference) ?? null,
   )
 
   const setBrowserTurnEndSound = (value: TurnEndSoundPreference | null) => {
     setBrowserTurnEndSoundSignal(value)
-    writeLocalStorage(KEY_TURN_END_SOUND, value)
+    updateBrowserPref('turnEndSound', value ?? undefined)
   }
 
   const [browserTurnEndSoundVolume, setBrowserTurnEndSoundVolumeSignal] = createSignal<number | null>(
-    (() => {
-      const v = readLocalStorage(KEY_TURN_END_SOUND_VOLUME)
-      return v !== null ? Number(v) : null
-    })(),
+    initialPrefs.turnEndSoundVolume ?? null,
   )
 
   const setBrowserTurnEndSoundVolume = (value: number | null) => {
     setBrowserTurnEndSoundVolumeSignal(value)
-    writeLocalStorage(KEY_TURN_END_SOUND_VOLUME, value !== null ? String(value) : null)
+    updateBrowserPref('turnEndSoundVolume', value ?? undefined)
   }
 
   const [browserDebugLogging, setBrowserDebugLoggingSignal] = createSignal<boolean | null>(
-    (() => {
-      const v = readLocalStorage(KEY_DEBUG_LOGGING)
-      if (v === null)
-        return null
-      return v === 'true'
-    })(),
+    initialPrefs.debugLogging ?? null,
   )
 
   const setBrowserDebugLogging = (value: boolean | null) => {
     setBrowserDebugLoggingSignal(value)
-    writeLocalStorage(KEY_DEBUG_LOGGING, value !== null ? String(value) : null)
+    updateBrowserPref('debugLogging', value ?? undefined)
   }
 
-  // --- Browser-only developer preferences ---
+  // --- Browser-only preferences ---
   const [showHiddenMessages, setShowHiddenMessagesSignal] = createSignal(
-    readLocalStorage(KEY_SHOW_HIDDEN_MESSAGES) === 'true',
+    initialPrefs.showHiddenMessages === true,
   )
   const setShowHiddenMessages = (value: boolean) => {
     setShowHiddenMessagesSignal(value)
-    writeLocalStorage(KEY_SHOW_HIDDEN_MESSAGES, value ? 'true' : null)
+    updateBrowserPref('showHiddenMessages', value || undefined)
+  }
+
+  const [enterKeyMode, setEnterKeyModeSignal] = createSignal<EnterKeyMode>(
+    initialPrefs.enterKeyMode ?? 'cmd-enter-sends',
+  )
+  const setEnterKeyMode = (value: EnterKeyMode) => {
+    setEnterKeyModeSignal(value)
+    updateBrowserPref('enterKeyMode', value)
   }
 
   // --- Resolved values (browser override → account default → hardcoded) ---
@@ -294,6 +289,8 @@ export const PreferencesProvider: ParentComponent = (props) => {
       debugLogging,
       showHiddenMessages,
       setShowHiddenMessages,
+      enterKeyMode,
+      setEnterKeyMode,
       uiFontCustomEnabled,
       monoFontCustomEnabled,
       uiFonts,
